+++ src/main/java/io/github/aimtone/tentacolous/annotations/ActionListener.java
@@ -1,0 +1,7 @@
+package io.github.aimtone.tentacolous.annotations;
+
+public enum ActionListener {
+    INSERT,
+    UPDATE,
+    DELETE
+}